backend/requirements.txt
fastapi[standard]
openai-agents
python-dotenv
loguru

